Add feature: File download and %20 escaping on upload
When uploading files with names that contain %20 in their names we need to escape it to empty space. Since filenames are passed via url string, it is impossible to tell whether the file contains an empty space ot the char sequence %20 in its name. So everything is converted to empty spaces when uploaded to the server.